feat: add upper bound for encrypted random integers
Add an upper bound for encrypted random integers by introducing a
separate precompile that does a modulo operation before trivially
encrypting the result. Generate encrypted random integers in the
[0, upperBound) range where upperBound is a power of 2. If the given
upperBound is not a power of 2, return an error. Reason for using
powers of 2 is potential performance benefit when using an FHE-based
random number generation in comparisson to an arbitrary upperBound.

Avoid potential out-of-bound indexing of input in `FheLib`. The start
index would be invalid if the actual input is exactly 4 bytes - in that
case, indexing would start from index 4 and that doesn't exist. Fix by
taking the min of 4 and len(input) - 1.

Add unit tests for encrypted random number generation with an
upperBound.