feature(decrypt): call KMS for decryption
IMPORTANT: the decrypt call should be replaced in most cases by cmux
It is still called in Governor example in fhevm repository.
To make tests passing we keep it for now.
In the future, it will be replaced by async decryption
For this early first version of KMS this is acceptable.