feat: remove protected storage/memory
This is the first phase of a process of using an ACL contract instead of
protected storage and memory in the fhEVM.

Instead of protected storage contracts, we now store all ciphertexts in
a single contract at address 94.

We no longer use protected memory for cross-contract calls.

fhEVM is NOT secure after this commit. Adding ACL checks will make it
secure again in a future commit.